refactor: split helpers into category modules
